Added Mobile Phone Property to all Layers
With this current commit adding Mobile Phone, user can edit his mobile phone and save it for thedelivery address. [Added to all layers]

With last pull request the new thing was inserted wherer instead of passing just the client object in EditProfile,  I'm passing the model view in which contains client object and the properties for the countries and cities.
With this ists easier to pass more variable/objects to a view.